Fixed misc compil warnings

+++ src/isomedia/meta.c
@@ -420,7 +420,7 @@ GF_Err gf_isom_set_meta_xml(GF_ISOFile *file, Bool root_meta, u32 track_num, cha
 	length = (u32) gf_ftell(xmlfile);
 	gf_fseek(xmlfile, 0, SEEK_SET);
 	xml->xml = (char*)gf_malloc(sizeof(unsigned char)*length);
-	bread = fread(xml->xml, 1, sizeof(unsigned char)*length, xmlfile);
+	bread =  (u32) fread(xml->xml, 1, sizeof(unsigned char)*length, xmlfile);
 	if (ferror(xmlfile) || (bread != length)) {
 		gf_free(xml->xml);
 		xml->xml = NULL;
